fix accuracy issue in timerefresh
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11520 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_main.c
@@ -2015,7 +2015,7 @@ For program optimization
 static void CL_TimeRefresh_f (void)
 {
 	int i;
-	float timestart, timedelta;
+	double timestart, timedelta;
 
 	r_refdef.scene.extraupdate = false;
